Squashed 'backend/' changes from 1ff8700..765a827
765a827 Merge pull request 'dev' (!167) from dev into test
1cdd7f1 Merge pull request 'feature/5131' (!166) from feature/5131 into dev
643e26b added other fields into table where sending details from user table refs #5131
0777a72 Merge pull request 'dev' (!165) from dev into test
8cbdcf6 Merge pull request 'bugfix/5134' (!164) from bugfix/5134 into dev
03a4d87 In the Tenant user role add the active from the user table and rename the avtive refs #5134
241c08d In the Tenant user responce remove tenant id and add user id and in the tenant user add the user id and change the name to the active refs #5134
e49da3f Merge branch 'dev' into feature/5131
54c6386 Tenant Dashboard card components refs #5131

git-subtree-dir: backend
git-subtree-split: 765a827c949b5e8d1bd186102d43cea1761e1da1

+++ src/main/java/com/payvance/erp_saas/core/repository/TenantUsageRepository.java
@@ -32,6 +32,7 @@ public interface TenantUsageRepository extends JpaRepository<TenantUsage, Long> 
      * Get companies count directly from companies_count column
      */
     @Query("SELECT tu.companiesCount FROM TenantUsage tu WHERE tu.tenantId = :tenantId")
-    Integer findCompaniesCountByTenantId(@Param("tenantId") Long tenantId);
+    Long findCompaniesCountByTenantId(@Param("tenantId") Long tenantId);
     boolean existsByTenantId(Long tenantId);
+    
 }
